[NVMCTRL] Updated nvmctrl_flash_read_write application for pic32cm_pl10_curiosity_nano
MH3C-101

+++ apps/nvmctrl/nvmctrl_flash_read_write/src/config/default/peripheral/nvmctrl/plib_nvmctrl.h
@@ -78,7 +78,7 @@
 /* NVMCTRL lock error */
 #define   NVMCTRL_ERROR_LOCK     ( 0x8U )
 
-typedef uint16_t NVMCTRL_ERROR;
+typedef uint32_t NVMCTRL_ERROR;
 
 
 typedef enum
